Feat edit card api

+++ src/main/java/com/codurance/retropolis/controllers/BoardController.java
@@ -21,7 +21,7 @@ public class BoardController extends BaseController {
   }
 
   @GetMapping(value = "/{id}")
-  public Board getBoard(@PathVariable Integer id) {
+  public Board getBoard(@PathVariable Long id) {
     return boardService.getBoard(id);
   }
 }
